backend/requirements.txt
# FastAPI Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.26.0

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis
redis==5.0.1

# Celery (Background Tasks)
celery==5.3.6

# Storage (S3/MinIO)
boto3==1.34.34

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
python-dotenv==1.0.1
